% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetSpecNames.R
\name{GetSpecNames}
\alias{GetSpecNames}
\title{Get Species official name}
\usage{
GetSpecNames(Species = NULL)
}
\arguments{
\item{Species}{Names of species to be queried. Case is ignored.}
}
\value{
list of species scientific name and species name
}
\description{
Check available species official scientific name
}
\examples{
\dontrun{
# Get scientific name for single or multiple species:
species_name <- GetSpecNames(c("mouse","human"))

# Or get a full list:
species_name <- GetSpecNames(Species = NULL)
}
}
