% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HomoHVG_Object.R
\docType{class}
\name{HomoHVG-class}
\alias{HomoHVG-class}
\title{Create a HomoHVG object}
\arguments{
\item{species1, species2}{Names of species in comparative analysis. Case is ignored}

\item{species1_mat, species2_mat}{Single cell expression matrix of species 1 or 2 with row as gene symbol/ID and column as sample ID}

\item{species1_mat_homo, species2_mat_homo}{Single cell expression matrix of species 1 or 2 with row as homologous gene name/ID and column as sample ID}

\item{spec1_HVG, spec2_HVG}{Homo-HVG list of species 1 or 2}

\item{Table_homo}{Table of one to one homologous genes list between species 1 and species 2}
}
\value{
HomoHVG object
}
\description{
HomoHVG object containing species matrics, homologous genes and variable genes information
}
\section{Slots}{

\describe{
\item{\code{Species}}{List of species name}

\item{\code{HomoHVG}}{List of Homo-HVGs}

\item{\code{Matrix_orig}}{List of original single cell matrix}

\item{\code{Matrix_homo}}{List of single cell matrix with homologous genes as row}

\item{\code{HVG_feature}}{List of HVG information table}

\item{\code{Table_homo}}{Table of homologous information}
}}

\concept{objects}
