% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HomologyHVG_Object.R
\docType{class}
\name{HomologyHVG-class}
\alias{HomologyHVG-class}
\title{Create a HomologyHVG object}
\arguments{
\item{RefSpec, QuySpec}{Names of reference(ref) and query(quy) species in comparative analysis. Case is ignored}

\item{RefSpec_mat, QuySpec_mat}{Single cell expression matrix of reference or query species with row as gene symbol/ID and column as sample ID}

\item{RefSpec_mat_homo, QuySpec_mat_homo}{Single cell expression matrix of reference or query species with row as homologous gene name/ID and column as sample ID}

\item{RefSpec_HVGs, QuySpec_HVGs}{Homology-HVG list of reference or query species}

\item{Table_homo}{Table of one to one homologous genes list between reference and query species}
}
\value{
HomologyHVG object
}
\description{
HomologyHVG object containing species matrics, homologous genes and variable genes information
}
\section{Slots}{

\describe{
\item{\code{Species}}{List of species name}

\item{\code{HomologyHVG}}{List of Homology-HVGs}

\item{\code{Matrix_orig}}{List of original single cell matrix}

\item{\code{Matrix_homo}}{List of single cell matrix with homologous genes as row}

\item{\code{HVG_feature}}{List of HVG information table}

\item{\code{Table_homo}}{Table of homologous information}
}}

\concept{objects}
