% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HomoHVG_Object.R
\name{HomoHVGObject}
\alias{HomoHVGObject}
\title{Title}
\usage{
HomoHVGObject(
  species1,
  species2,
  species1_mat,
  species2_mat,
  species1_mat_homo,
  species2_mat_homo,
  spec1_HVG,
  spec2_HVG,
  Table_homo
)
}
\arguments{
\item{Table_homo}{}
}
\description{
Title
}
