% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HomoSeeker.R
\name{HomoSeeker}
\alias{HomoSeeker}
\title{Title}
\usage{
HomoSeeker(
  species1,
  species2,
  species1_mat,
  species2_mat,
  species1_gene = "Gene_sym",
  species2_gene = "Gene_sym",
  method = "seurat",
  HVGs_method = "vst",
  verbose = T,
  mirror = NULL,
  version = NULL,
  host = "https://dec2021.archive.ensembl.org/"
)
}
\arguments{
\item{host}{}
}
\description{
Title
}
