% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HomoSelector.R
\name{HomoSelector}
\alias{HomoSelector}
\title{Title}
\usage{
HomoSelector(
  species1,
  species2,
  homotype = "ortholog_one2one",
  mirror = NULL,
  version = NULL,
  host = "https://dec2021.archive.ensembl.org/"
)
}
\arguments{
\item{host}{}
}
\description{
Title
}
