% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HVGSelector.R
\name{HVGSelector}
\alias{HVGSelector}
\title{Highly variable homologous genes identifier}
\usage{
HVGSelector(
  RefSpec,
  QuySpec,
  RefSpec_mat,
  QuySpec_mat,
  homo_mat = NULL,
  RefSpec_gene = "Gene_sym",
  QuySpec_gene = "Gene_sym",
  method = "seurat",
  HVGs_method = "vst",
  verbose = TRUE
)
}
\arguments{
\item{RefSpec, QuySpec}{Names of reference(ref) and query(quy) species in comparative analysis. Case is ignored}

\item{RefSpec_mat, QuySpec_mat}{Single cell expression matrix of reference or query species with row as gene symbol/ID and column as sample ID}

\item{homo_mat}{Table of one to one homologous genes table between species 1 and species 2. See HomoSelector() for further details.}

\item{RefSpec_gene, QuySpec_gene}{Type of gene name of Single cell expression matrix.
\itemize{
 \item{Gene_sym} \strong{:} Gene symbel (default).
 \item{Gene_id} \strong{:} Gene ID.
 }}

\item{method}{HVG selection method to be used. Currently Supported methods:
\itemize{
 \item{seurat}
 \item{scran}
 \item{scmap}
 }}

\item{HVGs_method}{if method = "Seurat", available methods:
\itemize{
 \item{vst(Default)} \strong{:} See Seurat::FindVariableFeatures().
 \item{scran} \strong{:} See Seurat::SCTransform().
 }}

\item{verbose}{Whether show calculation progress. Default is TRUE.}
}
\value{
A HomoHVG object with Highly variable homologous gene sets for both species
}
\description{
Identify highly variable genes using homologous gene set with HVG Methodologies
}
