% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HVGSelector.R
\name{HVGSelector}
\alias{HVGSelector}
\title{Title}
\usage{
HVGSelector(
  species1,
  species2,
  species1_mat,
  species2_mat,
  homo_mat = NULL,
  species1_gene = "Gene_sym",
  species2_gene = "Gene_sym",
  method = "seurat",
  HVGs_method = "vst",
  verbose = T
)
}
\arguments{
\item{verbose}{}
}
\description{
Title
}
